requirements.txt
langchain
langchain_core
langchain_community
python-dotenv
langchain_openai